refactor: 하위 컴포넌트에서 state 직접 관리하도록 수정
